Squashed initial commit of a .NET port of editorconfig-core
The current method for .NET relies on a managed c++ dll to wrap the core which is not portable to mono.

Test suite passes all but one tests which seems to pass on the commandline, will open a ticket later.

Still in need of buildscripts and nuget/chocolatey packaging and CI setup.

Would also love to port some of the tests to C# using System.IO.Abstractions but that might not make it in the initial release.